// src/@generated/body-detail/body-detail-min-aggregate.input.ts
import { Field } from '@nestjs/graphql';
import { InputType } from '@nestjs/graphql';

@InputType()
export class BodyDetailMinAggregateInput {
	@Field(() => Boolean, { nullable: true })
	id?: true;

	@Field(() => Boolean, { nullable: true })
	weight?: true;

	@Field(() => Boolean, { nullable: true })
	height?: true;

	@Field(() => Boolean, { nullable: true })
	age?: true;

	@Field(() => Boolean, { nullable: true })
	bmi?: true;

	@Field(() => Boolean, { nullable: true })
	bmr?: true;

	@Field(() => Boolean, { nullable: true })
	tdee?: true;

	@Field(() => Boolean, { nullable: true })
	userId?: true;

	@Field(() => Boolean, { nullable: true })
	createdAt?: true;

	@Field(() => Boolean, { nullable: true })
	updatedAt?: true;
}